Replace the fullscreen-SDL skip-raise heuristic with grabbing the server
SDL-1.2 in fullscreen mode draws to a override-redirect window it
places covering the entire screen. However, it still creates a normal
wm-controlled window that it will not paint to, and if we take too
long between mapping that window and raising it, then it can wind up
hiding the intended fullscreen window.

Rather than attempt to guess which windows are non-painted SDL
windows, just grab the server for the entire map+raise section.