Thread-safety improvements to the timer code
Implements the following fixes:

The value :CALL-IN-MAIN-THREAD is added to *FEATURES*, making the code
that checks for this much more clear.

Accesses to *TIMER-LIST* is now thread-safe on most platforms.

When adding a timer on platforms supporting CALL-IN-MAIN-THREAD, the
adding is scheduled in the main thread in order to make the timing more
accurate.

On platforms that supports CALL-IN-MAIN-THREAD, if the call is made in
the main thread itself, the function will be evaluated immediately.