Add the visual-pull-from-windowlist command.
Following on up #394, this new command implements a new way to cycle
through the list of open windows and select the appropriate one.

This version has no drawback that I'm aware of:

- If a window is on another frame, it won't be pulled.
- Windows are only pulled when necessary.
- Stack history of windows is properly kept in the group's list.

The real drawback of this solution is that it's using a hook that is
externalised, and it's the only one that is non-nil that I can
see. Should it use a separate, internal-only, hook, that select-menu
will also call?